data type to int32

+++ models/chart/compare_maps.py
@@ -63,7 +63,7 @@ def compare_maps(map1, map2, des, bitshift=3, stable=True, overwrite=False):
     # compare maps
     log.info('Comparing maps')
     try:
-        array3 = array1 * (10**bitshift) + array2
+        array3 = array1.astype(int32) * (10**bitshift) + array2.astype(int32)
         if stable:
             array3[array1==array2] = 0
         if geo['nodata'] != 'NA':
